Extract fill-color frame source
